etl: Refactor diagnostic report creation code
There is much overlap between the `etl redcap-det kiosk` and `etl
redcap-det uw-retrospectives` code for creating diagnostic reports to
add to a FHIR Bundle. Refactor this code to a shared module, accepting
a new argument, `create_device_result_observation_resource`, to
dynamically generate a list of observations with device-specific logic
to attach to the diagnostic reports.
